tests/phase_3_5/requirements.txt
matplotlib
numpy>=1.21.0
pytest>=7.0.0
pytest-cov>=3.0.0
PyPDF2>=3.0.0
pytesseract>=0.3.8
Pillow>=9.0.0
pdf2image>=1.16.0
opencv-python>=4.5.0
psutil>=5.9.0